# Orion/openapi.yaml
openapi: 3.0.3
info:
  title: Orion 用户服务 API
  version: 1.0.0
  description: |
    Orion 用户服务。认证由 Authentik OIDC 和 Gateway 负责，Orion 不提供注册、登录、刷新 Token、登出或修改密码接口。

    客户端只需向 Gateway 发送 Bearer Access Token。Gateway 验签后向 Orion 注入可信身份信息，Orion 使用 Authentik `sub` UUID 自动建档。
    所有接口和字段均带有中文说明，导入 Apifox 后可直接生成接口目录、请求模型和响应模型。
servers:
  - url: http://localhost:8091
    description: 本地 Gateway
...
security:
  - bearerAuth: []
paths:
...
  /v1/files/avatar/presign:
    post:
      tags: [头像]
      summary: 获取头像上传预签名 URL
      description: 只生成临时对象键和 PUT URL；RN 应直接向对象存储上传，不经过 Orion。支持 image/jpeg、image/png、image/webp、image/gif。
      operationId: presignAvatarUpload
      requestBody: { $ref: '#/components/requestBodies/AvatarPresign' }
      responses:
        '200': { $ref: '#/components/responses/PresignedUrlResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /v1/files/avatar/confirm:
    post:
      tags: [头像]
      summary: 确认头像上传
      description: 请求体必须是 objectKey 本身的 JSON 字符串，且必须属于当前用户的 avatar-staging/{userId}/ 命名空间；确认后转为正式 avatars/{userId}/ 对象。
      operationId: confirmAvatarUpload
      requestBody: { $ref: '#/components/requestBodies/AvatarConfirm' }
      responses:
        '200': { $ref: '#/components/responses/AvatarConfirmResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
  /v1/files/avatar:
    get:
      tags: [头像]
      summary: 获取当前用户头像访问 URL
      description: 返回短期有效的预签名 GET URL；未设置头像时返回 404。
      operationId: getAvatarUrl
      responses:
        '200': { $ref: '#/components/responses/AvatarUrlResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentik OIDC access_token。客户端不要伪造 X-Auth-* 或 X-Gateway-Source 请求头。
...
  requestBodies:
...
    AvatarPresign:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AvatarPresignRequest' }
    AvatarConfirm:
      required: true
      content:
        application/json:
          schema:
            type: string
            description: 'objectKey 本身的 JSON 字符串，不是 {objectKey: ...} 对象。'
            example: avatar-staging/550e8400-e29b-41d4-a716-446655440000/1723190400000_a1b2c3d4.webp
  responses:
    BadRequest:
      description: 请求参数无效或业务校验失败
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    Unauthorized:
      description: 缺少或无效的 Authentik Bearer Token
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    Forbidden:
      description: 邮箱未验证、业务账户已暂停/注销或对象不属于当前用户
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    NotFound:
      description: 用户、档案、业务记录或头像不存在
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
...
    ServerError:
      description: 服务器内部错误
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
...
    PresignedUrlResponse:
      description: 头像预签名上传信息
      content: { application/json: { schema: { $ref: '#/components/schemas/PresignedUrlResponse' } } }
    AvatarConfirmResponse:
      description: 头像确认结果
      content: { application/json: { schema: { $ref: '#/components/schemas/AvatarConfirmResponse' } } }
    AvatarUrlResponse:
      description: 当前头像访问 URL
      content: { application/json: { schema: { $ref: '#/components/schemas/AvatarUrlResponse' } } }
...
  schemas:
    ApiResponseBase:
      type: object
      required: [code, message, timestamp]
      properties:
        code: { type: integer, format: int32, description: HTTP 风格业务状态码, example: 200 }
        message: { type: string, description: 结果说明, example: Success }
        traceId: { type: string, nullable: true, description: 请求链路 ID }
        timestamp: { type: integer, format: int64, description: 服务端生成的毫秒时间戳, example: 1723190400000 }
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data: { nullable: true, description: 失败时固定为空 }
...
    PresignedUrlResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties: { data: { $ref: '#/components/schemas/PresignedUrl' } }
    AvatarConfirmResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties: { data: { $ref: '#/components/schemas/AvatarConfirm' } }
    AvatarUrlResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties: { data: { type: string, format: uri, description: 头像访问 URL } }
...
    AvatarPresignRequest:
      type: object
      required: [fileName, contentType]
      properties:
        fileName: { type: string, minLength: 1, description: 原始文件名 }
        contentType: { type: string, enum: [image/jpeg, image/png, image/webp, image/gif], description: 文件 MIME 类型 }
...
    PresignedUrl:
      type: object
      required: [uploadUrl, objectKey, expiresIn]
      properties:
        uploadUrl: { type: string, format: uri, description: 预签名 PUT URL }
        objectKey: { type: string, description: 临时对象键，确认时原样提交 }
        expiresIn: { type: integer, format: int64, description: URL 有效期，单位秒 }
    AvatarConfirm:
      type: object
      required: [avatarUrl]
      properties:
        avatarUrl: { type: string, format: uri, description: 最终头像访问 URL }